added:bwtwalk test
git-svn-id: http://svn.seqan.de/seqan/trunk@4876 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/tests/index/test_index_creation.h
@@ -94,14 +94,14 @@ namespace SEQAN_NAMESPACE_MAIN
         for (i = 0;  i < n;  i++)
 			if ((TSigned)SA[i] >= 0 && SA[i] < n) {
                 if (seen[SA[i]])
-                    printf("isPermutation: not unique %d->%d\n", i, SA[i]);
+                    printf("isPermutation: not unique %d->%d\n", (int)i, (int)SA[i]);
                 seen[SA[i]] = true;
             } else
-                printf("isPermutation: SA index out of range (n=%d) SA[%d]=%d\n", n, i, SA[i]);
+                printf("isPermutation: SA index out of range (n=%d) SA[%d]=%d\n", (int)n, (int)i, (int)SA[i]);
 
         for (i = 0;  i < n;  i++)
             if (!seen[i]) {
-                printf("isPermutation: not surjective %d empty\n", i);
+                printf("isPermutation: not surjective %d empty\n", (int)i);
                 delete[] seen;
                 return false;
             }
@@ -120,15 +120,15 @@ namespace SEQAN_NAMESPACE_MAIN
         for (i = 0;  i < n;  i++, ++SA)
             if (*SA >= 0 && *SA < n) {
                 if (seen[*SA])
-                    printf("isPermutation: not unique %d->%d\n", i, *SA);
+                    printf("isPermutation: not unique %d->%d\n", (int)i, (int)*SA);
                 seen[*SA] = true;
             } else
-                printf("isPermutation: SA index out of range (n=%d) SA[%d]=%d\n", n, i, *SA);
+                printf("isPermutation: SA index out of range (n=%d) SA[%d]=%d\n", (int)n, (int)i, (int)*SA);
         endRead(SA);
 
         for (i = 0;  i < n;  i++)
             if (!seen[i]) {
-                printf("isPermutation: not surjective %d empty\n", i);
+                printf("isPermutation: not surjective %d empty\n", (int)i);
                 delete[] seen;
                 return false;
             }
@@ -165,7 +165,7 @@ namespace SEQAN_NAMESPACE_MAIN
         TSize n = length(s);
 	    for(TSize i = 1; i < n; ++i) {
 		    if (!__sleq(begin(s) + SA[i-1], begin(s) + SA[i], n-SA[i-1], n-SA[i])) {
-			    printf("isSorted: sort error s_%d(SA[%d]) >= s_%d(SA[%d])\n",SA[i-1],i-1,SA[i],i);
+			    printf("isSorted: sort error s_%d(SA[%d]) >= s_%d(SA[%d])\n",(int)SA[i-1],(int)(i-1),(int)SA[i],(int)i);
 /*
 			String<unsigned, External<> > safile;
 			if (!open(safile,"error.sa")) printf("could not open ERROR.SA\n");
@@ -184,7 +184,7 @@ namespace SEQAN_NAMESPACE_MAIN
         TSize prev = *SA; ++SA;
 	    for(TSize i = 1; !eof(SA); ++SA, ++i) {
 		    if (!__sleq(begin(s) + prev, begin(s) + *SA, n-prev, n-*SA)) {
-                printf("isSorted: sort error s_%d(SA[%d]) >= s_%d(SA[%d])\n",prev,i-1,*SA,i);
+                printf("isSorted: sort error s_%d(SA[%d]) >= s_%d(SA[%d])\n",(int)prev,(int)(i-1),(int)*SA,(int)i);
 				endRead(SA);
 
 /*			String<unsigned, External<> > safile;
@@ -247,7 +247,7 @@ namespace SEQAN_NAMESPACE_MAIN
 	    TSize prev = *sa; ++sa;
 	    for(TSize i = 1; sa != end(SA); ++sa, ++lcp, ++i) {
 		    if (!__sleqLCP(begin(s) + prev, begin(s) + *sa, n-prev, n-*sa, *lcp)) {
-                printf("isLCP: sort error s_%d(%d) >= s_%d(%d)\n",i-1,prev,i,*sa);
+                printf("isLCP: sort error s_%d(%d) >= s_%d(%d)\n",(int)(i-1),(int)prev,(int)i,(int)*sa);
 			    return false;
 		    }
 		    prev = *sa;
@@ -282,12 +282,12 @@ namespace SEQAN_NAMESPACE_MAIN
     template <typename TLCP, typename TSufArray, typename TText>
     bool isLCPTable(TLCP &LCP, TSufArray &SA, TText const &s) {
         if (length(SA) != length(s)) {
-            printf("isLCPTable: length is bad: SA=%d, s=%d\n", length(SA), length(s));
+            printf("isLCPTable: length is bad: SA=%d, s=%d\n", (int)length(SA), (int)length(s));
             return false;
         }
         
         if (length(LCP) != length(s)) {
-            printf("isLCPTable: length is bad: LCP=%d, s=%d\n", length(LCP), length(s));
+            printf("isLCPTable: length is bad: LCP=%d, s=%d\n", (int)length(LCP), (int)length(s));
             return false;
         }
